refactor(gpu,lifecycle): #77 split gpu.py + extract _lifecycle.py _do_swap helper (< 300 LOC)
- Replace gpu.py module with gpu/ package:
  - gpu/vram.py: probe_free_vram_gib, VRAMSampler, VRAMMonitor (226 LOC)
  - gpu/kv.py: _QUANT_BITS, _quant_multiplier, kv_overhead_gib (88 LOC)
  - gpu/__init__.py: backward-compat re-exports (13 LOC)
- Extract _swap_drain_and_replace from _do_swap into nats/_swap.py (93 LOC)
  - _lifecycle.py drops from 339 to 290 LOC
- Remove both files from tools/file_exemptions.txt
- Update test patches to target llmcli.gpu.vram.subprocess.run
- All quality gates pass: ruff lint + typecheck, pytest 306/306, import-linter
