+++ src/llmcli/gpu/__init__.py
@@ -1,0 +1,13 @@
+"""GPU utilities for llmCLI — VRAM probing and KV-cache overhead estimation."""
+
+from __future__ import annotations
+
+from llmcli.gpu.kv import kv_overhead_gib
+from llmcli.gpu.vram import VRAMMonitor, VRAMSampler, probe_free_vram_gib
+
+__all__ = [
+    "probe_free_vram_gib",
+    "VRAMSampler",
+    "VRAMMonitor",
+    "kv_overhead_gib",
+]
