+++ tools/file_exemptions.txt
@@ -1,5 +1,3 @@
 # File size exemptions (≤300 lines per file)
 # Each entry must have a tracking issue.
 # Format: <path>  # <lines> lines — <issue> <description>
-src/llmcli/nats/_lifecycle.py  # 328 lines — DEBT:file-exemptions — #77 LifecycleMixin landed in #34/ADR-004 Slice 6; _do_swap drain+stop+start extraction deferred
-src/llmcli/gpu.py  # 311 lines — DEBT:file-exemptions — #77 VRAMSampler/VRAMMonitor + kv_overhead split into gpu/vram.py + gpu/kv.py deferred
